programming/requirements.txt
click==6.7
pkg-resources==0.0.0
pyserial==3.4
